+++ ClothesShop/Models/CartDtl.cs
@@ -9,7 +9,7 @@ public partial class CartDtl
 
     public string PdId { get; set; } = null!;
 
-    public double? CdtlQty { get; set; }
+    public int? CdtlQty { get; set; }
 
     public double? CdtlPrice { get; set; }
